+++ lib/src/plugins/mod.rs
@@ -1,3 +1,9 @@
+// Extendeded resources:
 pub mod invite;
+
+// Endpoints:
+// When adding an endpoint, add it to the list of endpoints in lib/src/endpoints.rs
+pub mod files;
 pub mod path;
+pub mod search;
 pub mod versioning;